Tighten get-card extractor boundaries
Apply the review findings that still held against the current tree.

- remove entity-empty module fallback during candidate selection
- split extractor helpers into candidate, position, and utility modules
- reject relative paths before building file URIs and normalise test
  callers to use absolute paths
- update the ExecPlan wording to remove the absolute path and the
  duplicated help-section sentence

The suggested wildcard removal in  was verified and left
unchanged because  is non-exhaustive.
